refactor(rel): extract scalar conversion adapters (#1178)
