+++ crates/graphforge-rel/src/expr.rs
@@ -5,6 +5,15 @@
 //! that can be consumed by operator lowering (#575, #576).
 
 mod aggregates;
+mod scalar_adapters;
+use scalar_adapters::{
+    CYPHER_TO_BOOLEAN, CYPHER_TO_FLOAT, CYPHER_TO_INTEGER, CYPHER_TO_STRING, render_temporal,
+    spatial_scalar,
+};
+#[cfg(test)]
+use scalar_adapters::{CypherToString, cypher_float_string};
+pub use scalar_adapters::{ir_literal_to_scalar, scalar_to_ir_literal};
+
 mod temporal_udfs;
 use temporal_udfs::{
     CYPHER_DATE_COMPONENT, CYPHER_DATE_PROJECT, CYPHER_DATE_TRUNCATE, CYPHER_DATETIME_PROJECT,
@@ -4702,334 +4711,6 @@ fn lower_literal(lit_val: &IrLiteral) -> DfExpr {
     lit(ir_literal_to_scalar(lit_val))
 }
 
-/// Convert an [`IrLiteral`] to a DataFusion [`ScalarValue`].
-///
-/// The single source of truth for the IR-literal → Arrow-scalar mapping, used
-/// both for literal expression lowering ([`lower_literal`]) and for binding
-/// query parameters to placeholder values (`$param` injection, #584).
-#[must_use]
-pub fn ir_literal_to_scalar(lit_val: &IrLiteral) -> ScalarValue {
-    match lit_val {
-        IrLiteral::Null => ScalarValue::Null,
-        IrLiteral::Bool(b) => ScalarValue::Boolean(Some(*b)),
-        IrLiteral::Int(n) => ScalarValue::Int64(Some(*n)),
-        IrLiteral::Float(f) => ScalarValue::Float64(Some(*f)),
-        IrLiteral::Str(s) => ScalarValue::Utf8(Some(s.clone())),
-        IrLiteral::Uuid(uuid) => ScalarValue::FixedSizeBinary(16, Some(uuid.to_vec())),
-        IrLiteral::Duration {
-            months,
-            days,
-            seconds,
-            nanos,
-        } => duration_scalar(Some(crate::temporal::DurationValue {
-            months: *months,
-            days: *days,
-            seconds: *seconds,
-            nanos: *nanos,
-        })),
-        IrLiteral::DateTime(us) => ScalarValue::TimestampMicrosecond(Some(*us), Some("UTC".into())),
-        IrLiteral::Date(days) => date_scalar(Some(*days)),
-        IrLiteral::LocalDateTime { days, nanos } => localdatetime_scalar(Some((*days, *nanos))),
-        IrLiteral::Time(nanos) => ScalarValue::Time64Nanosecond(Some(*nanos)),
-        IrLiteral::ZonedTime { nanos, offset } => time_scalar(Some((*nanos, *offset))),
-        IrLiteral::ZonedDateTime {
-            days,
-            nanos,
-            offset,
-            zone,
-        } => datetime_scalar(Some((*days, *nanos, *offset, zone.clone()))),
-        IrLiteral::Spatial(value) => spatial_property_scalar(value),
-        // A homogeneous list → a `ScalarValue::List` of the element scalars; the
-        // inner type is the first element's (re-typing untyped nulls to it so the
-        // array stays homogeneous, as the list-literal lowering does). (#1006)
-        IrLiteral::List(items) => {
-            let scalars: Vec<ScalarValue> = items.iter().map(ir_literal_to_scalar).collect();
-            let elem_type = scalars
-                .iter()
-                .find(|s| !s.is_null())
-                .map_or(DataType::Null, ScalarValue::data_type);
-            let typed: Vec<ScalarValue> = scalars
-                .iter()
-                .map(|s| {
-                    if matches!(s, ScalarValue::Null) {
-                        ScalarValue::try_from(&elem_type).unwrap_or(ScalarValue::Null)
-                    } else {
-                        s.clone()
-                    }
-                })
-                .collect();
-            ScalarValue::List(ScalarValue::new_list(&typed, &elem_type, true))
-        }
-        IrLiteral::Map(entries) => {
-            let scalars: Vec<(String, ScalarValue)> = entries
-                .iter()
-                .map(|(key, value)| (key.clone(), ir_literal_to_scalar(value)))
-                .collect();
-            const_map_scalar(&scalars).expect("IR map literal should lower to an Arrow struct")
-        }
-    }
-}
-
-fn spatial_scalar(value: &graphforge_core::SpatialValue) -> ScalarValue {
-    use datafusion::arrow::array::{Float64Array, StructArray};
-    use datafusion::arrow::buffer::NullBuffer;
-    use datafusion::arrow::datatypes::{Field, Fields};
-    use graphforge_core::SpatialCoordinates;
-
-    fn coordinate(value: [f64; 2]) -> ScalarValue {
-        let fields = Fields::from(vec![
-            Field::new("x", DataType::Float64, false),
-            Field::new("y", DataType::Float64, false),
-        ]);
-        ScalarValue::Struct(Arc::new(StructArray::new(
-            fields,
-            vec![
-                Arc::new(Float64Array::from(vec![value[0]])),
-                Arc::new(Float64Array::from(vec![value[1]])),
-            ],
-            Some(NullBuffer::from(vec![true])),
-        )))
-    }
-
-    fn list(values: &[ScalarValue]) -> ScalarValue {
-        let data_type = values
-            .first()
-            .map_or(DataType::Null, ScalarValue::data_type);
-        ScalarValue::List(ScalarValue::new_list(values, &data_type, false))
-    }
-
-    fn coordinates(values: &[[f64; 2]]) -> ScalarValue {
-        list(&values.iter().copied().map(coordinate).collect::<Vec<_>>())
-    }
-
-    match &value.coordinates {
-        SpatialCoordinates::Point(value) => coordinate(*value),
-        SpatialCoordinates::LineString(values) | SpatialCoordinates::MultiPoint(values) => {
-            coordinates(values)
-        }
-        SpatialCoordinates::Polygon(parts) | SpatialCoordinates::MultiLineString(parts) => list(
-            &parts
-                .iter()
-                .map(|part| coordinates(part))
-                .collect::<Vec<_>>(),
-        ),
-        SpatialCoordinates::MultiPolygon(polygons) => list(
-            &polygons
-                .iter()
-                .map(|polygon| {
-                    list(
-                        &polygon
-                            .iter()
-                            .map(|ring| coordinates(ring))
-                            .collect::<Vec<_>>(),
-                    )
-                })
-                .collect::<Vec<_>>(),
-        ),
-    }
-}
-
-fn spatial_property_scalar(value: &graphforge_core::SpatialValue) -> ScalarValue {
-    use datafusion::arrow::array::{StringArray, StructArray};
-    use datafusion::arrow::buffer::NullBuffer;
-    use datafusion::arrow::datatypes::{Field, Fields};
-    let fields = Fields::from(vec![Field::new(
-        "__graphforge_spatial_v1",
-        DataType::Utf8,
-        false,
-    )]);
-    ScalarValue::Struct(Arc::new(StructArray::new(
-        fields,
-        vec![Arc::new(StringArray::from(vec![
-            serde_json::to_string(value).expect("spatial value serialization is infallible"),
-        ]))],
-        Some(NullBuffer::from(vec![true])),
-    )))
-}
-
-/// Convert a DataFusion [`ScalarValue`] back to an [`IrLiteral`] for storage as
-/// a property value (#791 SET).
-///
-/// The inverse of [`ir_literal_to_scalar`], used by the SET execution node:
-/// after evaluating a value expression per row it gets a `ScalarValue` that must
-/// be stored as an `IrLiteral`. Numeric and temporal widths the writer does not
-/// have a dedicated literal for are normalised to the nearest `IrLiteral`
-/// (smaller ints → `Int`, `Float32` → `Float`, dates/times → `DateTime`). A
-/// `null` scalar (typed or untyped) → [`IrLiteral::Null`].
-///
-/// Lists and temporal structs are converted recursively. Maps, graph values,
-/// and other non-property shapes return an invalid-property-type error.
-///
-/// # Errors
-/// Returns [`LoweringError::InvalidType`] for a value that openCypher does not
-/// permit as a stored property.
-#[allow(
-    clippy::too_many_lines,
-    reason = "closed exhaustive property scalar conversion table"
-)]
-pub fn scalar_to_ir_literal(value: &ScalarValue) -> Result<IrLiteral, LoweringError> {
-    if let Some(decoded) =
-        decode_het_scalar(value).map_err(|error| LoweringError::InvalidType(error.to_string()))?
-    {
-        return scalar_to_ir_literal(&decoded);
-    }
-    // Any null (typed `Int64(None)` or untyped `Null`) stores as Cypher null.
-    if value.is_null() {
-        return Ok(IrLiteral::Null);
-    }
-    let lit = match value {
-        ScalarValue::Boolean(Some(b)) => IrLiteral::Bool(*b),
-        ScalarValue::Int8(Some(n)) => IrLiteral::Int(i64::from(*n)),
-        ScalarValue::Int16(Some(n)) => IrLiteral::Int(i64::from(*n)),
-        ScalarValue::Int32(Some(n)) => IrLiteral::Int(i64::from(*n)),
-        ScalarValue::Int64(Some(n)) => IrLiteral::Int(*n),
-        ScalarValue::UInt8(Some(n)) => IrLiteral::Int(i64::from(*n)),
-        ScalarValue::UInt16(Some(n)) => IrLiteral::Int(i64::from(*n)),
-        ScalarValue::UInt32(Some(n)) => IrLiteral::Int(i64::from(*n)),
-        ScalarValue::UInt64(Some(n)) => i64::try_from(*n).map(IrLiteral::Int).map_err(|_| {
-            LoweringError::UnsupportedExpr(format!("SET value {n} exceeds the i64 range"))
-        })?,
-        ScalarValue::Float32(Some(f)) => IrLiteral::Float(f64::from(*f)),
-        ScalarValue::Float64(Some(f)) => IrLiteral::Float(*f),
-        ScalarValue::Utf8(Some(s))
-        | ScalarValue::LargeUtf8(Some(s))
-        | ScalarValue::Utf8View(Some(s)) => IrLiteral::Str(s.clone()),
-        ScalarValue::FixedSizeBinary(16, Some(_)) => {
-            return Err(LoweringError::InvalidType(
-                "UUID values cannot be stored as graph properties".into(),
-            ));
-        }
-        // Flat native-Arrow duration widths carry no month/day part. Split each
-        // unit into whole seconds + non-negative nanos-of-second WITHOUT forming a
-        // `*1e9` total (which would overflow i64 for large native durations — the
-        // seconds field stores them directly). (#1011)
-        ScalarValue::DurationSecond(Some(s)) => duration_value_to_ir(dur_secs_nanos(*s, 0)),
-        ScalarValue::DurationMillisecond(Some(ms)) => duration_value_to_ir(dur_secs_nanos(
-            ms.div_euclid(1_000),
-            ms.rem_euclid(1_000) * 1_000_000,
-        )),
-        ScalarValue::DurationMicrosecond(Some(us)) => duration_value_to_ir(dur_secs_nanos(
-            us.div_euclid(1_000_000),
-            us.rem_euclid(1_000_000) * 1_000,
-        )),
-        ScalarValue::DurationNanosecond(Some(ns)) => {
-            duration_value_to_ir(crate::temporal::DurationValue::from_total_nanos(0, 0, *ns))
-        }
-        ScalarValue::TimestampMicrosecond(Some(us), _) => IrLiteral::DateTime(*us),
-        ScalarValue::TimestampSecond(Some(s), _) => IrLiteral::DateTime(s * 1_000_000),
-        ScalarValue::TimestampMillisecond(Some(ms), _) => IrLiteral::DateTime(ms * 1_000),
-        ScalarValue::TimestampNanosecond(Some(ns), _) => IrLiteral::DateTime(ns / 1_000),
-        // A date keeps its date identity (ADR 0009/0012): a `Struct{epoch_day}` of
-        // i64 days, not coerced to a `DateTime` — so it reads back and renders as a
-        // date.
-        ScalarValue::Struct(arr)
-            if arr.fields().len() == 1 && arr.fields()[0].name() == "__graphforge_spatial_v1" =>
-        {
-            let json = arr
-                .column(0)
-                .as_any()
-                .downcast_ref::<datafusion::arrow::array::StringArray>()
-                .ok_or_else(|| {
-                    LoweringError::InvalidType("malformed internal spatial value".into())
-                })?
-                .value(0);
-            IrLiteral::Spatial(serde_json::from_str(json).map_err(|_| {
-                LoweringError::InvalidType("malformed internal spatial value".into())
-            })?)
-        }
-        ScalarValue::Struct(arr) if is_date_struct(&DataType::Struct(arr.fields().clone())) => {
-            match date_struct_value(arr, 0) {
-                Some(days) => IrLiteral::Date(days),
-                None => IrLiteral::Null,
-            }
-        }
-        // A typed `duration` struct (#920) keeps its months/days/seconds/nanos model.
-        ScalarValue::Struct(arr) if is_duration_struct(&DataType::Struct(arr.fields().clone())) => {
-            match duration_struct_parts(arr, 0) {
-                Some(d) => duration_value_to_ir(d),
-                None => IrLiteral::Null,
-            }
-        }
-        // A typed `localdatetime` struct (#920): date-days + nanos-of-day.
-        ScalarValue::Struct(arr)
-            if is_localdatetime_struct(&DataType::Struct(arr.fields().clone())) =>
-        {
-            match localdatetime_struct_parts(arr, 0) {
-                Some((days, nanos)) => IrLiteral::LocalDateTime { days, nanos },
-                None => IrLiteral::Null,
-            }
-        }
-        // A typed `localtime` (#920): nanoseconds-of-day, no zone.
-        ScalarValue::Time64Nanosecond(Some(n)) => IrLiteral::Time(*n),
-        // A typed `time` struct (#920): time-of-day + UTC offset.
-        ScalarValue::Struct(arr) if is_time_struct(&DataType::Struct(arr.fields().clone())) => {
-            match time_struct_parts(arr, 0) {
-                Some((nanos, offset)) => IrLiteral::ZonedTime { nanos, offset },
-                None => IrLiteral::Null,
-            }
-        }
-        // A typed `datetime` struct (#920): date + time + offset + named zone.
-        ScalarValue::Struct(arr) if is_datetime_struct(&DataType::Struct(arr.fields().clone())) => {
-            match datetime_struct_parts(arr, 0) {
-                Some((days, nanos, offset, zone)) => IrLiteral::ZonedDateTime {
-                    days,
-                    nanos,
-                    offset,
-                    zone,
-                },
-                None => IrLiteral::Null,
-            }
-        }
-        // A homogeneous list (#1006): recurse element-wise (a null element →
-        // `IrLiteral::Null`; an element type the gate rejects propagates its
-        // error). Single-row `List`/`LargeList` scalars carry the elements at
-        // row 0.
-        ScalarValue::List(arr) => list_scalar_to_ir_literal(&arr.value(0))?,
-        ScalarValue::LargeList(arr) => list_scalar_to_ir_literal(&arr.value(0))?,
-        other => {
-            // Plain maps, graph structs, and any width not handled above are
-            // invalid openCypher property values.
-            return Err(LoweringError::InvalidType(format!(
-                "invalid property type for SET value: {:?}",
-                other.data_type()
-            )));
-        }
-    };
-    Ok(lit)
-}
-
-/// Build an [`IrLiteral::List`] from a list scalar's element array, recursing
-/// through [`scalar_to_ir_literal`] per element. (#1006)
-fn list_scalar_to_ir_literal(
-    elems: &datafusion::arrow::array::ArrayRef,
-) -> Result<IrLiteral, LoweringError> {
-    let mut items = Vec::with_capacity(elems.len());
-    for j in 0..elems.len() {
-        let ev = ScalarValue::try_from_array(elems, j).map_err(|e| {
-            LoweringError::UnsupportedExpr(format!("list element is not a scalar value: {e}"))
-        })?;
-        items.push(scalar_to_ir_literal(&ev)?);
-    }
-    Ok(IrLiteral::List(items))
-}
-
-/// Canonicalise a literal-string temporal-constructor argument. Returns `None`
-/// when the constructor doesn't take a string, or the string isn't a form the
-/// `temporal` module recognises (the caller then falls back to the runtime
-/// path). (#599)
-fn render_temporal(name: &str, s: &str) -> Option<String> {
-    use crate::temporal;
-    match name {
-        "date" => temporal::render_date(s),
-        "localtime" => temporal::render_local_time(s),
-        "time" => temporal::render_time(s),
-        "localdatetime" => temporal::render_local_date_time(s),
-        "datetime" => temporal::render_date_time(s),
-        "duration" => temporal::render_duration(s),
-        _ => None,
-    }
-}
-
 /// Look up a Cypher built-in function name and produce the DataFusion call.
 ///
 /// Returns `None` if the name is not in the built-in table.
@@ -5522,219 +5203,12 @@ fn cypher_slice(list: DfExpr, start: Option<DfExpr>, end: Option<DfExpr>) -> DfE
 // Cypher conversion UDFs
 // ---------------------------------------------------------------------------
 
-static CYPHER_TO_INTEGER: LazyLock<ScalarUDF> = LazyLock::new(|| {
-    ScalarUDF::new_from_impl(CypherConversion::new(CypherConversionKind::Integer))
-});
-static CYPHER_TO_FLOAT: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherConversion::new(CypherConversionKind::Float)));
-static CYPHER_TO_BOOLEAN: LazyLock<ScalarUDF> = LazyLock::new(|| {
-    ScalarUDF::new_from_impl(CypherConversion::new(CypherConversionKind::Boolean))
-});
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-enum CypherConversionKind {
-    Integer,
-    Float,
-    Boolean,
-}
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherConversion {
-    kind: CypherConversionKind,
-    signature: Signature,
-}
-
-impl CypherConversion {
-    fn new(kind: CypherConversionKind) -> Self {
-        Self {
-            kind,
-            signature: Signature::any(1, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherConversion {
-    fn name(&self) -> &'static str {
-        match self.kind {
-            CypherConversionKind::Integer => "cypher_to_integer",
-            CypherConversionKind::Float => "cypher_to_float",
-            CypherConversionKind::Boolean => "cypher_to_boolean",
-        }
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(match self.kind {
-            CypherConversionKind::Integer => DataType::Int64,
-            CypherConversionKind::Float => DataType::Float64,
-            CypherConversionKind::Boolean => DataType::Boolean,
-        })
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use datafusion::arrow::array::{BooleanArray, Float64Array, Int64Array};
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let array = args.args[0].to_array(args.number_rows)?;
-        match self.kind {
-            CypherConversionKind::Integer => {
-                let out: datafusion::error::Result<Int64Array> = (0..array.len())
-                    .map(|i| {
-                        let value = decoded_scalar_at(&array, i)?;
-                        to_cypher_integer(&value)
-                    })
-                    .collect();
-                Ok(ColumnarValue::Array(std::sync::Arc::new(out?)))
-            }
-            CypherConversionKind::Float => {
-                let out: datafusion::error::Result<Float64Array> = (0..array.len())
-                    .map(|i| {
-                        let value = decoded_scalar_at(&array, i)?;
-                        to_cypher_float(&value)
-                    })
-                    .collect();
-                Ok(ColumnarValue::Array(std::sync::Arc::new(out?)))
-            }
-            CypherConversionKind::Boolean => {
-                let out: datafusion::error::Result<BooleanArray> = (0..array.len())
-                    .map(|i| {
-                        let value = decoded_scalar_at(&array, i)?;
-                        to_cypher_boolean(&value)
-                    })
-                    .collect();
-                Ok(ColumnarValue::Array(std::sync::Arc::new(out?)))
-            }
-        }
-    }
-}
-
 fn decoded_scalar_at(
     array: &datafusion::arrow::array::ArrayRef,
     row: usize,
 ) -> datafusion::error::Result<ScalarValue> {
     let value = ScalarValue::try_from_array(array, row)?;
     Ok(unwrap_het(value))
-}
-
-fn conversion_type_error(fn_name: &str, value: &ScalarValue) -> datafusion::error::DataFusionError {
-    datafusion::error::DataFusionError::Execution(format!(
-        "{fn_name}() cannot convert value of type {:?}",
-        value.data_type()
-    ))
-}
-
-fn to_cypher_integer(value: &ScalarValue) -> datafusion::error::Result<Option<i64>> {
-    if value.is_null() {
-        return Ok(None);
-    }
-    if let Some(i) = scalar_as_i128(value) {
-        return i64::try_from(i)
-            .map(Some)
-            .map_err(|_| conversion_type_error("toInteger", value));
-    }
-    match value {
-        ScalarValue::Float32(Some(f)) => Ok(trunc_float_to_i64(f64::from(*f))),
-        ScalarValue::Float64(Some(f)) => Ok(trunc_float_to_i64(*f)),
-        ScalarValue::Utf8(Some(s)) | ScalarValue::LargeUtf8(Some(s)) => Ok(s
-            .parse::<f64>()
-            .ok()
-            .filter(|f| f.is_finite())
-            .and_then(trunc_float_to_i64)),
-        _ => Err(conversion_type_error("toInteger", value)),
-    }
-}
-
-#[allow(
-    clippy::cast_possible_truncation,
-    clippy::cast_precision_loss,
-    reason = "openCypher toInteger truncates finite floating values toward zero"
-)]
-fn trunc_float_to_i64(f: f64) -> Option<i64> {
-    if !f.is_finite() {
-        return None;
-    }
-    let truncated = f.trunc();
-    if truncated < i64::MIN as f64 || truncated > i64::MAX as f64 {
-        return None;
-    }
-    Some(truncated as i64)
-}
-
-fn to_cypher_float(value: &ScalarValue) -> datafusion::error::Result<Option<f64>> {
-    if value.is_null() {
-        return Ok(None);
-    }
-    if let Some(f) = scalar_as_f64(value) {
-        return Ok(Some(f));
-    }
-    match value {
-        ScalarValue::Utf8(Some(s)) | ScalarValue::LargeUtf8(Some(s)) => {
-            Ok(s.parse::<f64>().ok().filter(|f| f.is_finite()))
-        }
-        _ => Err(conversion_type_error("toFloat", value)),
-    }
-}
-
-fn to_cypher_boolean(value: &ScalarValue) -> datafusion::error::Result<Option<bool>> {
-    if value.is_null() {
-        return Ok(None);
-    }
-    match value {
-        ScalarValue::Boolean(Some(b)) => Ok(Some(*b)),
-        ScalarValue::Utf8(Some(s)) | ScalarValue::LargeUtf8(Some(s)) => match s.as_str() {
-            "true" => Ok(Some(true)),
-            "false" => Ok(Some(false)),
-            _ => Ok(None),
-        },
-        _ => Err(conversion_type_error("toBoolean", value)),
-    }
-}
-
-fn cypher_float_string(f: f64) -> String {
-    if f == 0.0 {
-        "0.0".to_owned()
-    } else if f.is_nan() {
-        "NaN".to_owned()
-    } else if f.is_infinite() {
-        if f.is_sign_positive() {
-            "Infinity".to_owned()
-        } else {
-            "-Infinity".to_owned()
-        }
-    } else {
-        let mut s = f.to_string();
-        if !s.contains('.') && !s.contains('e') && !s.contains('E') {
-            s.push_str(".0");
-        }
-        s
-    }
-}
-
-fn to_cypher_string(value: &ScalarValue) -> datafusion::error::Result<Option<String>> {
-    if value.is_null() {
-        return Ok(None);
-    }
-    match value {
-        ScalarValue::Int8(Some(n)) => Ok(Some(n.to_string())),
-        ScalarValue::Int16(Some(n)) => Ok(Some(n.to_string())),
-        ScalarValue::Int32(Some(n)) => Ok(Some(n.to_string())),
-        ScalarValue::Int64(Some(n)) => Ok(Some(n.to_string())),
-        ScalarValue::UInt8(Some(n)) => Ok(Some(n.to_string())),
-        ScalarValue::UInt16(Some(n)) => Ok(Some(n.to_string())),
-        ScalarValue::UInt32(Some(n)) => Ok(Some(n.to_string())),
-        ScalarValue::UInt64(Some(n)) => Ok(Some(n.to_string())),
-        ScalarValue::Float32(Some(f)) => Ok(Some(cypher_float_string(f64::from(*f)))),
-        ScalarValue::Float64(Some(f)) => Ok(Some(cypher_float_string(*f))),
-        ScalarValue::Boolean(Some(b)) => Ok(Some(b.to_string())),
-        ScalarValue::Utf8(Some(s)) | ScalarValue::LargeUtf8(Some(s)) => Ok(Some(s.clone())),
-        _ => Err(conversion_type_error("toString", value)),
-    }
 }
 
 // ---------------------------------------------------------------------------
@@ -7624,118 +7098,6 @@ impl ScalarUDFImpl for CypherListComp {
 // cypher_to_string UDF
 // ---------------------------------------------------------------------------
 
-/// `toString(x)`: a typed temporal value renders to its canonical openCypher
-/// string (a plain `cast` to `Utf8` would emit a fixed, untrimmed form for
-/// `Time64`, and outright fail for the temporal structs). Handles `Date32`/
-/// `Time64`/`localdatetime`/`time`; every other type — including `datetime`,
-/// which is still a `Utf8` value until its migration — falls back to the same
-/// `Utf8` cast as before, so non-temporal `toString` behaviour is unchanged.
-/// (ADR 0009)
-static CYPHER_TO_STRING: LazyLock<ScalarUDF> =
-    LazyLock::new(|| ScalarUDF::new_from_impl(CypherToString::new()));
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct CypherToString {
-    signature: Signature,
-}
-
-impl CypherToString {
-    fn new() -> Self {
-        Self {
-            signature: Signature::any(1, Volatility::Immutable),
-        }
-    }
-}
-
-impl ScalarUDFImpl for CypherToString {
-    fn name(&self) -> &'static str {
-        "cypher_to_string"
-    }
-
-    fn signature(&self) -> &Signature {
-        &self.signature
-    }
-
-    fn return_type(&self, _arg_types: &[DataType]) -> datafusion::error::Result<DataType> {
-        Ok(DataType::Utf8)
-    }
-
-    fn invoke_with_args(
-        &self,
-        args: ScalarFunctionArgs,
-    ) -> datafusion::error::Result<ColumnarValue> {
-        use crate::temporal::{
-            format_date, render_localdatetime, render_localtime_nanos, render_time_value,
-        };
-        use datafusion::arrow::array::{Array, StringArray, StructArray, Time64NanosecondArray};
-        use datafusion::arrow::compute::cast;
-        use datafusion::arrow::datatypes::TimeUnit;
-
-        validate_heterogeneous_arguments(&args.args)?;
-        let arr = args.args[0].to_array(args.number_rows)?;
-        let rows = arr.len();
-        // Render one canonical string per row from a closure, preserving nulls.
-        let render = |f: &dyn Fn(usize) -> Option<String>| -> ColumnarValue {
-            let out: StringArray = (0..rows)
-                .map(|i| if arr.is_null(i) { None } else { f(i) })
-                .collect();
-            ColumnarValue::Array(std::sync::Arc::new(out))
-        };
-
-        let result = match arr.data_type() {
-            DataType::Struct(_) if is_date_struct(arr.data_type()) => {
-                let s = arr.as_any().downcast_ref::<StructArray>().unwrap();
-                render(&|i| date_struct_value(s, i).map(format_date))
-            }
-            DataType::Time64(TimeUnit::Nanosecond) => {
-                let a = arr
-                    .as_any()
-                    .downcast_ref::<Time64NanosecondArray>()
-                    .unwrap();
-                render(&|i| Some(render_localtime_nanos(a.value(i))))
-            }
-            DataType::Struct(_) if is_localdatetime_struct(arr.data_type()) => {
-                let s = arr.as_any().downcast_ref::<StructArray>().unwrap();
-                render(&|i| {
-                    localdatetime_struct_parts(s, i).map(|(d, n)| render_localdatetime(d, n))
-                })
-            }
-            DataType::Struct(_) if is_time_struct(arr.data_type()) => {
-                let s = arr.as_any().downcast_ref::<StructArray>().unwrap();
-                render(&|i| time_struct_parts(s, i).map(|(n, o)| render_time_value(n, o)))
-            }
-            DataType::Struct(_) if is_datetime_struct(arr.data_type()) => {
-                let s = arr.as_any().downcast_ref::<StructArray>().unwrap();
-                render(&|i| {
-                    datetime_struct_parts(s, i).map(|(d, n, o, z)| {
-                        crate::temporal::render_datetime_value(d, n, o, z.as_deref())
-                    })
-                })
-            }
-            DataType::Struct(_) if is_duration_struct(arr.data_type()) => {
-                let s = arr.as_any().downcast_ref::<StructArray>().unwrap();
-                render(&|i| {
-                    duration_struct_parts(s, i).map(|d| crate::temporal::render_duration_value(&d))
-                })
-            }
-            DataType::Struct(_) if is_het_struct_type(Some(arr.data_type())) => {
-                let out: datafusion::error::Result<StringArray> = (0..rows)
-                    .map(|i| {
-                        let value = decoded_scalar_at(&arr, i)?;
-                        to_cypher_string(&value)
-                    })
-                    .collect();
-                ColumnarValue::Array(std::sync::Arc::new(out?))
-            }
-            // Non-temporal (or non-temporal struct): the original `Utf8` cast.
-            _ => ColumnarValue::Array(
-                cast(&arr, &DataType::Utf8).map_err(datafusion::error::DataFusionError::from)?,
-            ),
-        };
-        Ok(result)
-    }
-}
-
 // ---------------------------------------------------------------------------
 // cypher_path_nodes UDF
 // ---------------------------------------------------------------------------
@@ -9129,76 +8491,6 @@ mod tests {
     }
 
     #[test]
-    fn cypher_conversions_decode_tagged_values() {
-        use datafusion::arrow::array::{Array, ArrayRef, Float64Array, Int64Array, StringArray};
-        use datafusion::arrow::datatypes::Field;
-        use datafusion::config::ConfigOptions;
-        use datafusion::scalar::ScalarValue as S;
-        use std::sync::Arc;
-
-        let values: ArrayRef = Arc::new(
-            build_het_struct(
-                &[
-                    S::Int64(Some(2)),
-                    S::Float64(Some(2.9)),
-                    S::Utf8(Some("foo".to_owned())),
-                ],
-                0,
-            )
-            .unwrap(),
-        );
-        let invoke = |kind: CypherConversionKind| {
-            let udf = CypherConversion::new(kind);
-            let return_type = udf.return_type(&[values.data_type().clone()]).unwrap();
-            let args = ScalarFunctionArgs {
-                args: vec![ColumnarValue::Array(Arc::clone(&values))],
-                arg_fields: vec![Arc::new(Field::new("v", values.data_type().clone(), true))],
-                number_rows: values.len(),
-                return_field: Arc::new(Field::new("out", return_type, true)),
-                config_options: Arc::new(ConfigOptions::default()),
-            };
-            match udf.invoke_with_args(args).unwrap() {
-                ColumnarValue::Array(a) => a,
-                ColumnarValue::Scalar(s) => s.to_array_of_size(values.len()).unwrap(),
-            }
-        };
-
-        let ints = invoke(CypherConversionKind::Integer);
-        let ints = ints.as_any().downcast_ref::<Int64Array>().unwrap();
-        assert_eq!(ints.value(0), 2);
-        assert_eq!(ints.value(1), 2);
-        assert!(ints.is_null(2));
-
-        let floats = invoke(CypherConversionKind::Float);
-        let floats = floats.as_any().downcast_ref::<Float64Array>().unwrap();
-        assert_eq!(floats.value(0), 2.0);
-        assert_eq!(floats.value(1), 2.9);
-        assert!(floats.is_null(2));
-
-        let strings = match CYPHER_TO_STRING
-            .invoke_with_args(ScalarFunctionArgs {
-                args: vec![ColumnarValue::Array(values)],
-                arg_fields: vec![Arc::new(Field::new(
-                    "v",
-                    DataType::Struct(het_fields(0)),
-                    true,
-                ))],
-                number_rows: 3,
-                return_field: Arc::new(Field::new("out", DataType::Utf8, true)),
-                config_options: Arc::new(ConfigOptions::default()),
-            })
-            .unwrap()
-        {
-            ColumnarValue::Array(a) => a,
-            ColumnarValue::Scalar(s) => s.to_array_of_size(3).unwrap(),
-        };
-        let strings = strings.as_any().downcast_ref::<StringArray>().unwrap();
-        assert_eq!(strings.value(0), "2");
-        assert_eq!(strings.value(1), "2.9");
-        assert_eq!(strings.value(2), "foo");
-    }
-
-    #[test]
     fn het_list_map_roundtrip_and_order() {
         use datafusion::scalar::ScalarValue as S;
         // A plain map is recognised; a typed temporal struct is NOT (#1005).
@@ -10286,106 +9578,6 @@ mod tests {
     // -----------------------------------------------------------------------
 
     #[test]
-    fn scalar_to_ir_literal_round_trips_each_kind() {
-        // Each IrLiteral → ScalarValue → IrLiteral is identity for the canonical
-        // widths `ir_literal_to_scalar` emits.
-        for lit in [
-            IrLiteral::Bool(true),
-            IrLiteral::Int(42),
-            IrLiteral::Float(1.5),
-            IrLiteral::Str("hi".into()),
-            IrLiteral::Duration {
-                months: 14,
-                days: 3,
-                seconds: 5,
-                nanos: 1000,
-            },
-            IrLiteral::DateTime(1_700_000_000_000_000),
-        ] {
-            let scalar = ir_literal_to_scalar(&lit);
-            assert_eq!(scalar_to_ir_literal(&scalar).unwrap(), lit);
-        }
-    }
-
-    #[test]
-    fn scalar_to_ir_literal_rejects_graph_identity_as_a_property_value() {
-        let scalar = ir_literal_to_scalar(&IrLiteral::Uuid([0x42; 16]));
-        assert!(matches!(
-            scalar_to_ir_literal(&scalar),
-            Err(LoweringError::InvalidType(message))
-                if message == "UUID values cannot be stored as graph properties"
-        ));
-    }
-
-    #[test]
-    fn scalar_to_ir_literal_null_variants_map_to_null() {
-        assert_eq!(
-            scalar_to_ir_literal(&ScalarValue::Null).unwrap(),
-            IrLiteral::Null
-        );
-        assert_eq!(
-            scalar_to_ir_literal(&ScalarValue::Int64(None)).unwrap(),
-            IrLiteral::Null
-        );
-    }
-
-    #[test]
-    fn scalar_to_ir_literal_widens_smaller_ints() {
-        assert_eq!(
-            scalar_to_ir_literal(&ScalarValue::Int32(Some(7))).unwrap(),
-            IrLiteral::Int(7)
-        );
-        assert_eq!(
-            scalar_to_ir_literal(&ScalarValue::UInt8(Some(255))).unwrap(),
-            IrLiteral::Int(255)
-        );
-    }
-
-    #[test]
-    fn scalar_to_ir_literal_normalizes_all_native_widths_and_rejects_overflow() {
-        let cases = [
-            (ScalarValue::Int8(Some(-8)), IrLiteral::Int(-8)),
-            (ScalarValue::Int16(Some(-16)), IrLiteral::Int(-16)),
-            (ScalarValue::UInt16(Some(16)), IrLiteral::Int(16)),
-            (ScalarValue::UInt32(Some(32)), IrLiteral::Int(32)),
-            (ScalarValue::UInt64(Some(64)), IrLiteral::Int(64)),
-            (ScalarValue::Float32(Some(1.25)), IrLiteral::Float(1.25)),
-            (
-                ScalarValue::LargeUtf8(Some("large".into())),
-                IrLiteral::Str("large".into()),
-            ),
-            (
-                ScalarValue::Utf8View(Some("view".into())),
-                IrLiteral::Str("view".into()),
-            ),
-            (
-                ScalarValue::TimestampSecond(Some(2), None),
-                IrLiteral::DateTime(2_000_000),
-            ),
-            (
-                ScalarValue::TimestampMillisecond(Some(3), None),
-                IrLiteral::DateTime(3_000),
-            ),
-            (
-                ScalarValue::TimestampNanosecond(Some(4_000), None),
-                IrLiteral::DateTime(4),
-            ),
-            (ScalarValue::Time64Nanosecond(Some(5)), IrLiteral::Time(5)),
-        ];
-        for (scalar, expected) in cases {
-            assert_eq!(scalar_to_ir_literal(&scalar).unwrap(), expected);
-        }
-        assert!(matches!(
-            scalar_to_ir_literal(&ScalarValue::UInt64(Some(u64::MAX))),
-            Err(LoweringError::UnsupportedExpr(message)) if message.contains("exceeds the i64 range")
-        ));
-        assert!(matches!(
-            scalar_to_ir_literal(&ScalarValue::Binary(Some(vec![1, 2]))),
-            Err(LoweringError::InvalidType(message)) if message.contains("invalid property type")
-        ));
-    }
-
-    #[test]
     fn dynamic_access_helpers_cover_null_bounds_types_and_schema_errors() {
         use datafusion::arrow::datatypes::{Field, Fields};
 
@@ -10537,19 +9729,6 @@ mod tests {
                 .contains("dynamic map/property access key must be a string"),
             "{error}"
         );
-    }
-
-    #[test]
-    fn scalar_to_ir_literal_round_trips_a_list() {
-        // A homogeneous list now stores (#1006): scalar List → IrLiteral::List
-        // and back, element-wise — including a list of typed temporals.
-        for lit in [
-            IrLiteral::List(vec![IrLiteral::Int(1), IrLiteral::Int(2)]),
-            IrLiteral::List(vec![IrLiteral::Date(5428), IrLiteral::Date(5429)]),
-        ] {
-            let scalar = ir_literal_to_scalar(&lit);
-            assert_eq!(scalar_to_ir_literal(&scalar).unwrap(), lit);
-        }
     }
 
     #[test]
@@ -10918,108 +10097,6 @@ mod tests {
     }
 
     #[test]
-    fn scalar_conversion_helpers_exhaust_every_numeric_width_null_and_error_contract() {
-        let integers = [
-            (ScalarValue::Int8(Some(-8)), -8_i64),
-            (ScalarValue::Int16(Some(-16)), -16),
-            (ScalarValue::Int32(Some(-32)), -32),
-            (ScalarValue::Int64(Some(-64)), -64),
-            (ScalarValue::UInt8(Some(8)), 8),
-            (ScalarValue::UInt16(Some(16)), 16),
-            (ScalarValue::UInt32(Some(32)), 32),
-            (ScalarValue::UInt64(Some(64)), 64),
-        ];
-        for (value, expected) in &integers {
-            assert_eq!(scalar_as_i128(value), Some(i128::from(*expected)));
-            assert_eq!(scalar_as_f64(value), Some(*expected as f64));
-            assert_eq!(to_cypher_integer(value).unwrap(), Some(*expected));
-            assert_eq!(to_cypher_float(value).unwrap(), Some(*expected as f64));
-            assert_eq!(to_cypher_string(value).unwrap(), Some(expected.to_string()));
-        }
-
-        for (value, integer, float, text) in [
-            (
-                ScalarValue::Float32(Some(12.75)),
-                Some(12),
-                Some(12.75),
-                Some("12.75".to_owned()),
-            ),
-            (
-                ScalarValue::Float64(Some(-12.75)),
-                Some(-12),
-                Some(-12.75),
-                Some("-12.75".to_owned()),
-            ),
-            (
-                ScalarValue::Utf8(Some("42.9".into())),
-                Some(42),
-                Some(42.9),
-                Some("42.9".to_owned()),
-            ),
-            (
-                ScalarValue::LargeUtf8(Some("-3".into())),
-                Some(-3),
-                Some(-3.0),
-                Some("-3".to_owned()),
-            ),
-        ] {
-            assert_eq!(to_cypher_integer(&value).unwrap(), integer);
-            assert_eq!(to_cypher_float(&value).unwrap(), float);
-            assert_eq!(to_cypher_string(&value).unwrap(), text);
-        }
-
-        for null in [
-            ScalarValue::Null,
-            ScalarValue::Int64(None),
-            ScalarValue::Float64(None),
-            ScalarValue::Utf8(None),
-            ScalarValue::Boolean(None),
-        ] {
-            assert_eq!(to_cypher_integer(&null).unwrap(), None);
-            assert_eq!(to_cypher_float(&null).unwrap(), None);
-            assert_eq!(to_cypher_boolean(&null).unwrap(), None);
-            assert_eq!(to_cypher_string(&null).unwrap(), None);
-        }
-
-        for invalid_float in [f64::NAN, f64::INFINITY, f64::NEG_INFINITY, f64::MAX] {
-            assert_eq!(trunc_float_to_i64(invalid_float), None);
-        }
-        assert_eq!(trunc_float_to_i64(-9.99), Some(-9));
-        for invalid_text in ["", "not-a-number", "NaN", "inf"] {
-            let value = ScalarValue::Utf8(Some(invalid_text.into()));
-            assert_eq!(to_cypher_integer(&value).unwrap(), None);
-            assert_eq!(to_cypher_float(&value).unwrap(), None);
-        }
-        assert_eq!(
-            to_cypher_boolean(&ScalarValue::Boolean(Some(true))).unwrap(),
-            Some(true)
-        );
-        assert_eq!(
-            to_cypher_boolean(&ScalarValue::Utf8(Some("true".into()))).unwrap(),
-            Some(true)
-        );
-        assert_eq!(
-            to_cypher_boolean(&ScalarValue::LargeUtf8(Some("false".into()))).unwrap(),
-            Some(false)
-        );
-        assert_eq!(
-            to_cypher_boolean(&ScalarValue::Utf8(Some("TRUE".into()))).unwrap(),
-            None
-        );
-
-        for invalid in [
-            ScalarValue::Boolean(Some(true)),
-            ScalarValue::Binary(Some(vec![1])),
-        ] {
-            assert!(to_cypher_integer(&invalid).is_err());
-            assert!(to_cypher_float(&invalid).is_err());
-        }
-        assert!(to_cypher_boolean(&ScalarValue::Int64(Some(1))).is_err());
-        assert!(to_cypher_string(&ScalarValue::Binary(Some(vec![1]))).is_err());
-        assert!(to_cypher_integer(&ScalarValue::UInt64(Some(u64::MAX))).is_err());
-    }
-
-    #[test]
     fn canonical_float_strings_and_scalar_range_arguments_cover_boundaries() {
         for (value, expected) in [
             (0.0, "0.0"),
@@ -11058,102 +10135,6 @@ mod tests {
                 .to_string()
                 .contains("must be an integer")
         );
-    }
-
-    #[test]
-    fn temporal_literal_render_dispatch_and_ir_scalar_round_trip_matrix() {
-        for (name, input) in [
-            ("date", "2024-02-29"),
-            ("localtime", "12:34:56"),
-            ("time", "12:34:56+01:00"),
-            ("localdatetime", "2024-02-29T12:34:56"),
-            ("datetime", "2024-02-29T12:34:56Z"),
-            ("duration", "P1M2DT3S"),
-        ] {
-            assert!(render_temporal(name, input).is_some(), "{name}({input})");
-        }
-        assert_eq!(render_temporal("unknown", "2024-01-01"), None);
-        assert_eq!(render_temporal("date", "not-a-date"), None);
-
-        let literals = [
-            IrLiteral::Null,
-            IrLiteral::Bool(true),
-            IrLiteral::Int(-7),
-            IrLiteral::Float(1.25),
-            IrLiteral::Str("value".into()),
-            IrLiteral::Duration {
-                months: 1,
-                days: 2,
-                seconds: 3,
-                nanos: 4,
-            },
-            IrLiteral::DateTime(123),
-            IrLiteral::Date(20_000),
-            IrLiteral::LocalDateTime {
-                days: 20_000,
-                nanos: 123,
-            },
-            IrLiteral::Time(456),
-            IrLiteral::ZonedTime {
-                nanos: 789,
-                offset: 3_600,
-            },
-            IrLiteral::ZonedDateTime {
-                days: 20_000,
-                nanos: 999,
-                offset: -3_600,
-                zone: Some("America/Denver".into()),
-            },
-            IrLiteral::List(vec![IrLiteral::Int(1), IrLiteral::Null]),
-            IrLiteral::Map(vec![("answer".into(), IrLiteral::Int(42))]),
-        ];
-        for literal in literals {
-            let scalar = ir_literal_to_scalar(&literal);
-            if !matches!(literal, IrLiteral::Map(_)) {
-                assert_eq!(scalar_to_ir_literal(&scalar).unwrap(), literal);
-            }
-        }
-
-        for (scalar, expected) in [
-            (
-                ScalarValue::DurationSecond(Some(-2)),
-                IrLiteral::Duration {
-                    months: 0,
-                    days: 0,
-                    seconds: -2,
-                    nanos: 0,
-                },
-            ),
-            (
-                ScalarValue::DurationMillisecond(Some(-1)),
-                IrLiteral::Duration {
-                    months: 0,
-                    days: 0,
-                    seconds: -1,
-                    nanos: 999_000_000,
-                },
-            ),
-            (
-                ScalarValue::DurationMicrosecond(Some(-1)),
-                IrLiteral::Duration {
-                    months: 0,
-                    days: 0,
-                    seconds: -1,
-                    nanos: 999_999_000,
-                },
-            ),
-            (
-                ScalarValue::DurationNanosecond(Some(-1)),
-                IrLiteral::Duration {
-                    months: 0,
-                    days: 0,
-                    seconds: -1,
-                    nanos: 999_999_999,
-                },
-            ),
-        ] {
-            assert_eq!(scalar_to_ir_literal(&scalar).unwrap(), expected);
-        }
     }
 
     #[test]
